convert to long

+++ src/scip/tree.c
@@ -2949,7 +2949,7 @@ void treeFindSwitchForks(
    /* if not already found, continue searching the LP defining fork; it cannot be deeper than the common fork */
    if( lpfork == NULL )
    {
-      if( tree->focuslpfork != NULL && (int)(tree->focuslpfork->depth) > fork->depth )
+      if( tree->focuslpfork != NULL && tree->focuslpfork->depth > fork->depth )
       {
          /* focuslpfork is not on the same active path as the new node: we have to continue searching */
          lpfork = fork;
@@ -2967,7 +2967,7 @@ void treeFindSwitchForks(
          /* focuslpfork is on the same active path as the new node: old and new node have the same lpfork */
          lpfork = tree->focuslpfork;
       }
-      assert(lpfork == NULL || (int)(lpfork->depth) <= fork->depth);
+      assert(lpfork == NULL || lpfork->depth <= fork->depth);
       assert(lpfork == NULL || lpfork->active);
    }
    assert(lpfork == NULL
@@ -2981,7 +2981,7 @@ void treeFindSwitchForks(
     */
    if( lpstatefork == NULL )
    {
-      if( tree->focuslpstatefork != NULL && (int)(tree->focuslpstatefork->depth) > fork->depth )
+      if( tree->focuslpstatefork != NULL && tree->focuslpstatefork->depth > fork->depth )
       {
          /* focuslpstatefork is not on the same active path as the new node: we have to continue searching */
          if( lpfork != NULL && lpfork->depth < fork->depth )
@@ -3001,7 +3001,7 @@ void treeFindSwitchForks(
          /* focuslpstatefork is on the same active path as the new node: old and new node have the same lpstatefork */
          lpstatefork = tree->focuslpstatefork;
       }
-      assert(lpstatefork == NULL || (int)(lpstatefork->depth) <= fork->depth);
+      assert(lpstatefork == NULL || lpstatefork->depth <= fork->depth);
       assert(lpstatefork == NULL || lpstatefork->active);
    }
    assert(lpstatefork == NULL
@@ -3015,7 +3015,7 @@ void treeFindSwitchForks(
     */
    if( subroot == NULL )
    {
-      if( tree->focussubroot != NULL && (int)(tree->focussubroot->depth) > fork->depth )
+      if( tree->focussubroot != NULL && tree->focussubroot->depth > fork->depth )
       {
          /* focussubroot is not on the same active path as the new node: we have to continue searching */
          if( lpstatefork != NULL && lpstatefork->depth < fork->depth )
@@ -4414,7 +4414,7 @@ SCIP_RETCODE SCIPnodeFocus(
    {
       /* we are in a different subtree, or no valid LP fork exists: the LP is completely incorrect */
       assert(subroot == NULL || !subroot->active
-         || (tree->focussubroot != NULL && (int)(tree->focussubroot->depth) > subroot->depth));
+         || (tree->focussubroot != NULL && tree->focussubroot->depth > subroot->depth));
       tree->correctlpdepth = -1;
    }
 
@@ -8274,7 +8274,7 @@ SCIP_Bool SCIPtreeIsPathComplete(
    assert(tree->pathlen == 0 || tree->focusnode != NULL);
    assert(tree->pathlen >= 2 || !SCIPtreeProbing(tree));
    assert(tree->pathlen == 0 || tree->path[tree->pathlen-1] != NULL);
-   assert(tree->pathlen == 0 || tree->path[tree->pathlen-1]->depth == tree->pathlen-1);
+   assert(tree->pathlen == 0 || (long)(tree->path[tree->pathlen-1]->depth) == tree->pathlen-1);
    assert(tree->focusnode == NULL || (int)tree->focusnode->depth >= tree->pathlen
       || tree->path[tree->focusnode->depth] == tree->focusnode);
 
@@ -8317,7 +8317,7 @@ SCIP_NODE* SCIPtreeGetFocusNode(
    assert(tree->pathlen == 0 || tree->focusnode != NULL);
    assert(tree->pathlen >= 2 || !SCIPtreeProbing(tree));
    assert(tree->pathlen == 0 || tree->path[tree->pathlen-1] != NULL);
-   assert(tree->pathlen == 0 || tree->path[tree->pathlen-1]->depth == tree->pathlen-1);
+   assert(tree->pathlen == 0 || (long)(tree->path[tree->pathlen-1]->depth) == tree->pathlen-1);
    assert(tree->focusnode == NULL || (int)tree->focusnode->depth >= tree->pathlen
       || tree->path[tree->focusnode->depth] == tree->focusnode);
 
@@ -8334,7 +8334,7 @@ int SCIPtreeGetFocusDepth(
    assert(tree->pathlen == 0 || tree->focusnode != NULL);
    assert(tree->pathlen >= 2 || !SCIPtreeProbing(tree));
    assert(tree->pathlen == 0 || tree->path[tree->pathlen-1] != NULL);
-   assert(tree->pathlen == 0 || tree->path[tree->pathlen-1]->depth == tree->pathlen-1);
+   assert(tree->pathlen == 0 || (long)(tree->path[tree->pathlen-1]->depth) == tree->pathlen-1);
    assert(tree->focusnode == NULL || (int)tree->focusnode->depth >= tree->pathlen
       || tree->path[tree->focusnode->depth] == tree->focusnode);
 
@@ -8392,7 +8392,7 @@ SCIP_NODE* SCIPtreeGetCurrentNode(
    assert(tree->pathlen == 0 || tree->focusnode != NULL);
    assert(tree->pathlen >= 2 || !SCIPtreeProbing(tree));
    assert(tree->pathlen == 0 || tree->path[tree->pathlen-1] != NULL);
-   assert(tree->pathlen == 0 || tree->path[tree->pathlen-1]->depth == tree->pathlen-1);
+   assert(tree->pathlen == 0 || (long)(tree->path[tree->pathlen-1]->depth) == tree->pathlen-1);
    assert(tree->focusnode == NULL || (int)tree->focusnode->depth >= tree->pathlen
       || tree->path[tree->focusnode->depth] == tree->focusnode);
 
@@ -8409,7 +8409,7 @@ int SCIPtreeGetCurrentDepth(
    assert(tree->pathlen == 0 || tree->focusnode != NULL);
    assert(tree->pathlen >= 2 || !SCIPtreeProbing(tree));
    assert(tree->pathlen == 0 || tree->path[tree->pathlen-1] != NULL);
-   assert(tree->pathlen == 0 || tree->path[tree->pathlen-1]->depth == tree->pathlen-1);
+   assert(tree->pathlen == 0 || (long)(tree->path[tree->pathlen-1]->depth) == tree->pathlen-1);
    assert(tree->focusnode == NULL || (int)tree->focusnode->depth >= tree->pathlen
       || tree->path[tree->focusnode->depth] == tree->focusnode);
